Add jamierocks' toy browser
This is my entry :^) I didn't get to spend as much time building this
as I would have liked, but it was good fun none-the-less.

While canvas' rich text functionality was very helpful for quickly
getting something going, I would manually write to a canvas if I were
to pursue this further.