Update readme to mention supporting ladybird as a browser
It's fixed now! Thanks to CxByte on discord for troubleshooting it. The
issue was my DNS was configured to use ipv6 but nginx wasn't actually
running ipv6. I removed ipv6 from the dns settings temporarily and I'll
figure out the nginx config later on

I checked 1.1.1.1, and 8.8.8.8 and both seemed to have refreshed, but
the DNS settings may be cached regionly for about another day.